Relocate python library and add docs (#377)
* move frontend/python to calyx-py

* add docs on the calyx-py library
