+++ calyx-py/README.md
@@ -5,7 +5,7 @@ To use the FuTIL AST and utilities:
 1. Install [flit](https://flit.readthedocs.io/en/latest/#install).
 2. Install the `futil` package:
 ```
-  $ cd frontends/python
+  $ cd calyx-py/
   $ flit install --symlink
 ```
