+++ docs/compiler-docs.md
@@ -1,0 +1,5 @@
+# Compiler Documentation
+
+The source code documentation for the compiler can be [found here][comp].
+
+[comp]: https://capra.cs.cornell.edu/calyx/doc/calyx/
